removed unused test script. Added utility script to add id and ref tags for topics that conform to /campus/building/device/optional sub-device/point format topic names
